# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Schemes: [http]",
    "title": "Captain ATC API",
    "version": "0.0.0"
  },
  "basePath": "/",
  "paths": {
    "/airspace": {
      "post": {
        "description": "Creates an airspace\nCreates an isolated environment for flights and formations to be provisioned.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "airspace"
        ],
        "operationId": "CreateAirspace",
        "parameters": [
          {
            "description": "Human-readable name for this airspace",
            "name": "Airspace",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "HumanName",
                "NetName"
              ],
              "properties": {
                "HumanName": {
                  "description": "Human-readable name for airspace.",
                  "type": "string"
                },
                "NetName": {
                  "description": "Name used for DNS name building, and internal queries against the state database.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request processed",
            "schema": {
              "type": "object",
              "properties": {
                "HumanName": {
                  "description": "Human-readable name for airspace.",
                  "type": "string"
                },
                "ID": {
                  "description": "Unique airspace ID in state database.",
                  "type": "integer"
                },
                "NetName": {
                  "description": "Name used for DNS name building, and internal queries against the state database.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error, possibly a database or validation error."
          }
        }
      }
    },
    "/airspace/{id}": {
      "get": {
        "description": "Gets an airspace stored in the state database. Does not auto-populate the Flight field.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "airspace"
        ],
        "summary": "Get an airspace managed by this ATC instance.",
        "operationId": "GetAirspace",
        "responses": {
          "200": {
            "description": "Request processed",
            "schema": {
              "type": "object",
              "properties": {
                "HumanName": {
                  "description": "Human-readable name for airspace.",
                  "type": "string"
                },
                "ID": {
                  "description": "Unique airspace ID in state database.",
                  "type": "integer"
                },
                "NetName": {
                  "description": "Name used for DNS name building, and internal queries against the state database.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error, possibly a database error."
          }
        }
      },
      "put": {
        "description": "Updates an airspace\nUpdates the properties of an airspace. Note that only the HumanName can be changed after creation.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "airspace"
        ],
        "operationId": "UpdateAirspace",
        "parameters": [
          {
            "description": "Unique ID of the airspace to get.",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "description": "Human-readable name for this airspace",
            "name": "Airspace",
            "in": "body",
            "schema": {
              "type": "object",
              "required": [
                "HumanName"
              ],
              "properties": {
                "HumanName": {
                  "description": "Human-readable name for airspace.",
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request processed",
            "schema": {
              "type": "object",
              "properties": {
                "HumanName": {
                  "description": "Human-readable name for airspace.",
                  "type": "string"
                },
                "ID": {
                  "description": "Unique airspace ID in state database.",
                  "type": "integer"
                },
                "NetName": {
                  "description": "Name used for DNS name building, and internal queries against the state database.",
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error, possibly a database or validation error."
          }
        }
      },
      "delete": {
        "description": "Gets an airspace stored in the state database. Does not auto-populate the Flight field.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "airspace"
        ],
        "summary": "Get an airspace managed by this ATC instance.",
        "operationId": "DeleteAirspace",
        "parameters": [
          {
            "description": "Unique ID of the airspace to get.",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request processed"
          },
          "500": {
            "description": "Internal server error, possibly a database error."
          }
        }
      }
    },
    "/airspaces": {
      "get": {
        "description": "Gets a list of all airspaces stored in the state database. Does not auto-populate the Flight field.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "airspace"
        ],
        "summary": "Get all airspaces managed by this ATC instance.",
        "operationId": "GetAirspaces",
        "responses": {
          "200": {
            "description": "Request processed",
            "schema": {
              "type": "array",
              "items": {
                "properties": {
                  "HumanName": {
                    "description": "Human-readable name for airspace.",
                    "type": "string"
                  },
                  "ID": {
                    "description": "Unique airspace ID in state database.",
                    "type": "integer"
                  },
                  "NetName": {
                    "description": "Name used for DNS name building, and internal queries against the state database.",
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error, possibly a database error."
          }
        }
      }
    },
    "/flight/{id}": {
      "delete": {
        "description": "Deletes a flight, and any dependent formations and planes.",
        "produces": [
          "application/json"
        ],
        "tags": [
          "flight"
        ],
        "summary": "Deletes a flight from the state database.",
        "operationId": "DeleteFlight",
        "parameters": [
          {
            "description": "Unique ID of the flight to delete.",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request processed"
          },
          "500": {
            "description": "Internal server error, possibly a database error."
          }
        }
      }
    }
  }
}